requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
pymupdf
langchain-text-splitters
sentence_transformers
langchain_huggingface
faiss-cpu
streamlit
langchain_groq